requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyMySQL==1.1.2
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
